// server/src/api/types/User.ts
export type User = {
  id: number;
  login: string;
  url: string;
  type: string;
  name: string;
  company: string;
  blog: string;
  location: string;
  email: string;
  html_url: string;
  avatar_url: string;
  public_repos: number;
  public_gists: number;
  followers: number;
  following: number;
};